@@ -31,9 +31,8 @@
     "intersection-observer": "^0.10.0",
     "netlify-cms-app": "^2.12.15",
     "prop-types": "^15.7.2",
-    "react": "^16.12.0",
-    "react-countdown": "^2.3.1",
-    "react-dom": "^16.12.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-helmet": "^6.0.0",
     "react-html-parser": "^2.0.2",
     "react-laag": "1.7.3",
