@@ -2,7 +2,7 @@ import React, { useCallback } from 'react';
 import axios from 'axios';
 import { useMutation } from '@apollo/client';
 import { loader } from 'graphql.macro';
-import { useHistory } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import { useAuth0 } from '@auth0/auth0-react';
 
 import LoginForm, {
@@ -16,7 +16,7 @@ const CREATE_USER = loader('../graphql/ReallyCreateUser.graphql');
 
 function Signup(): JSX.Element {
   const [createUser] = useMutation(CREATE_USER);
-  const history = useHistory();
+  const navigate = useNavigate();
   const { loginWithRedirect } = useAuth0();
 
   const submit = useCallback(
@@ -52,14 +52,14 @@ function Signup(): JSX.Element {
                       })
                       .then(resp => {
                         resolve();
-                        history.push(`/verify-email?email=${email}`);
+                        navigate(`/verify-email?email=${email}`);
                       })
                       .catch(err => {
                         reject(err);
                       });
                   } else {
                     resolve();
-                    history.push(`/verify-email?email=${email}`);
+                    navigate(`/verify-email?email=${email}`);
                   }
                 })
                 .catch(err => {
@@ -70,7 +70,7 @@ function Signup(): JSX.Element {
         );
       });
     },
-    [createUser, history],
+    [createUser, navigate],
   );
 
   return (
