@@ -1,10 +1,7 @@
 import React from 'react';
 import { makeStyles, DefaultTheme as Theme } from '@mui/styles';
 import MuiCheckbox from '@mui/material/Checkbox';
-import {
-  fieldToCheckbox,
-  CheckboxProps as MuiCheckboxProps,
-} from 'formik-mui';
+import { fieldToCheckbox, CheckboxProps as MuiCheckboxProps } from 'formik-mui';
 
 import CheckedIcon from '../icons/CheckedIcon';
 import UncheckedIcon from '../icons/UncheckedIcon';
