@@ -151,7 +151,7 @@ function IndividualModal({
         >
           {({ submitForm, isValid, isSubmitting }) => {
             return (
-              <Form translate="yes">
+              <Form>
                 <OnBoardingCard className={styles.card}>
                   <Field
                     className={styles.field}
